Move summary tab and update imports
