+++ src/views/workflow-summary/workflow-summary-details/workflow-summary-details.styles.ts
