+++ src/views/workflow-summary/workflow-summary-diagnostics-banner/workflow-summary-diagnostics-banner.types.ts
